Adding training .py files with CNTK and Tensorflow
The .py files are scripts used to efficiently train sequence-to-sequence models on >100million pair data sets. 

Tensorflow code is a LSTM-based ranker with LSTMs as encoders, and a winner-take-all (top-1) ranking objective utilizing a softmax function. The code logic initiates, runs a parameter server architecture and deal with communications automatically. 
CNTK code applies distributed training wrapped in the 'trainer' code. Data parallel is used on minibatches to aggregate gradients.